chore: abstract cfd password modal
