+++ packages/cfd/src/Containers/cfd-password-modal/__tests__/cfd-password-modal.spec.js
@@ -4,7 +4,7 @@ import { createBrowserHistory } from 'history';
 import { WS, getErrorMessages, validPassword, Jurisdiction } from '@deriv/shared';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import CFDPasswordModal from '../cfd-password-modal';
-import CFDProviders from '../../cfd-providers';
+import CFDProviders from '../../../cfd-providers';
 import { mockStore } from '@deriv/stores';
 
 jest.mock('@deriv/account', () => ({
